% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-functions.R
\name{get_bootstrap_prop_means}
\alias{get_bootstrap_prop_means}
\title{run bootstrap and return list of bootstraps with their corresponding mean}
\usage{
get_bootstrap_prop_means(original_dat, ntimes = 1000)
}
\arguments{
\item{original_dat}{A tibble or data frame with \code{prop} variable}

\item{ntimes}{A number}
}
\value{
\code{ntimes} bootstrap resamples and means of \code{original_dat}
}
\description{
run bootstrap and return list of bootstraps with their corresponding mean
}
\examples{
get_bootstrap_prop_means(original_dat = data.frame(x = rnorm(10), prop=rbeta(10,2,2)))
get_bootstrap_prop_means(original_dat = data.frame(x = rexp(5), prop= rbeta(5,1,1), names=letters(seq(from = 1, to=5)))
}
